Support PHP 5.3, take 2
Apparently PHP 5.3 considers an anonymous function defined inside the
class to be outside the class's scope, so any variables you want to use
inside the function have to be public. *sigh*

The alternative option would be to make the anonymous callback function
a private method of the class, which would allow you to use self:: and
keep the $widgets property private, but separating it like that feels
a bit messy to me.

Change-Id: I4b8921b6d22589d47356031a426a02da9f173b51