Convert Widgets to extension registration
This patch is converting the Widgets extension to extension registration.
The complicated namespace logic used in the old initilaization file is
no longer needed with extension registration and MW 1.30+, so it's removed.
Remaining code that was defined in Widgets.php before, but is too complicated
to be convertable to extension.json, is moved to a new WidgetInitializer class.

Bug: T191721
Change-Id: I36a7e199f29dcd7226f2ae37e5fc7a76a985f53a