Make forward compatible with Smarty v4.
These changes make the extension compatible with both smarty v4 and v3

Change timestamps to be unix timestamps. The smarty docs seem to imply
that that is what smarty expects. Previously they were MW formatted.

This changes the default php handling to passthru. All other methods
were removed in smarty v3. Passthru just means echo the tags, so it
should be just as secure as the remove method.

Bug: T330406
Change-Id: I55e334b7bf64dee5d6c920b9b9dfcafab6e3a481